strands-agents/enterprise-ai/sales-data-prep/requirements.txt
faker
boto3
python-dotenv
tabulate
pandas
SQLAlchemy
psycopg2-binary